feat: FoodItemsPanel component has been added
The FoodItemsPanel is one of the main components in our welcoming
dashboard. This panel is made up of a grid of FoodItemCard(s).
These cards are the components that hold the food item's data.
So far this component includes 3 rows of 3 cards and a search bar.
Current state is just for display, it is not connected to the back-end
services yet.